Refactoring

+++ src/main/java/com/test/task/foodordering/repository/MainRepo.java
@@ -5,6 +5,5 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
 @Repository
-public interface MainRepo extends JpaRepository<Main, Integer> {
-
+public interface MainRepo extends JpaRepository<Main, Long> {
 }
